+++ pymc3/plots/__init__.py
@@ -1,4 +1,4 @@
-#   Copyright 2020 The PyMC Developers
+#   Copyright 2021 The PyMC Developers
 #
 #   Licensed under the Apache License, Version 2.0 (the "License");
 #   you may not use this file except in compliance with the License.
@@ -12,12 +12,11 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
-"""PyMC3 Plotting.
+"""Alias for the `plots` submodule from ArviZ.
 
-Plots are delegated to the `ArviZ <https://arviz-devs.github.io/arviz/>`_ library, a general purpose library for
-exploratory analysis of Bayesian models. For more details, see https://arviz-devs.github.io/arviz/.
-
-Only `plot_posterior_predictive_glm` is kept in the PyMC code base for now, but it will move to ArviZ once the latter adds features for regression plots.
+Plots are delegated to the ArviZ library, a general purpose library for
+"exploratory analysis of Bayesian models."
+See https://arviz-devs.github.io/arviz/ for details on plots.
 """
 import functools
 import sys
@@ -25,6 +24,49 @@ import warnings
 
 import arviz as az
 
+# Makes this module as identical to arviz.plots as possible
+for attr in az.plots.__all__:
+    obj = getattr(az.plots, attr)
+    if not attr.startswith("__"):
+        setattr(sys.modules[__name__], attr, obj)
+
+
+def alias_deprecation(func, alias: str):
+    original = func.__name__
+
+    @functools.wraps(func)
+    def wrapped(*args, **kwargs):
+        raise DeprecationWarning(
+            f"The function `{alias}` from PyMC3 was an alias for `{original}` from ArviZ. "
+            "It was removed in PyMC3 4.0. "
+            f"Switch to `pymc3.{original}` or `arviz.{original}`."
+        )
+
+    return wrapped
+
+
+# Aliases of ArviZ functions
+autocorrplot = alias_deprecation(az.plot_autocorr, alias="autocorrplot")
+forestplot = alias_deprecation(az.plot_forest, alias="forestplot")
+kdeplot = alias_deprecation(az.plot_kde, alias="kdeplot")
+energyplot = alias_deprecation(az.plot_energy, alias="energyplot")
+densityplot = alias_deprecation(az.plot_density, alias="densityplot")
+pairplot = alias_deprecation(az.plot_pair, alias="pairplot")
+traceplot = alias_deprecation(az.plot_trace, alias="traceplot")
+compareplot = alias_deprecation(az.plot_compare, alias="compareplot")
+
+
 from pymc3.plots.posteriorplot import plot_posterior_predictive_glm
 
-__all__ = ["plot_posterior_predictive_glm"]
+__all__ = tuple(az.plots.__all__) + (
+    "autocorrplot",
+    "compareplot",
+    "forestplot",
+    "kdeplot",
+    "plot_posterior",
+    "traceplot",
+    "energyplot",
+    "densityplot",
+    "pairplot",
+    "plot_posterior_predictive_glm",
+)